refactor(backend): consolidate orchestrator files 5→4 (#4048)
- Remove Orchestrator/EnhancedOrchestrator backward-compat aliases from orchestrator.py
- Fold enhanced_orchestration/orchestrator.py into enhanced_orchestration/__init__.py
- Delete enhanced_orchestration/orchestrator.py (directory kept for other modules)
- Migrate all call-sites to ConsolidatedOrchestrator directly
