+++ autobot-backend/orchestration/plan_steps.e2e_test.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 sys.path.append(str(Path(__file__).parent))
 
-from orchestrator import Orchestrator, TaskComplexity
+from orchestrator import ConsolidatedOrchestrator as Orchestrator, TaskComplexity
 
 
 def test_plan_workflow_steps():
