+++ autobot-backend/orchestration/workflow_execution.e2e_test.py
@@ -10,7 +10,7 @@ from pathlib import Path
 # Add AutoBot to Python path
 sys.path.append(str(Path(__file__).parent))
 
-from orchestrator import Orchestrator
+from orchestrator import ConsolidatedOrchestrator as Orchestrator
 
 
 async def test_workflow_execution():
