+++ autobot-backend/orchestration/workflow_orchestration.e2e_test.py
@@ -11,7 +11,7 @@ import sys
 # Add src to Python path
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), "src"))
 
-from orchestrator import Orchestrator
+from orchestrator import ConsolidatedOrchestrator as Orchestrator
 
 
 async def test_workflow_classification():
